appimage: add some system libraries to the bundled dependencies
When an appimage is built with linuxdeploy,
some of the dependent libs are somehow not included in the bundle
(even though they are not listed in the 'excludelist').

they might be updated to newer API version at a host system,
(and libplacebo was indeed updated recently),
which makes the mpv appimage to fail.